+++ java-native-game/src/main/java/com/calebe/engine/modules/GameObjectModule.java
@@ -1,6 +1,7 @@
 package com.calebe.engine.modules;
 
 import com.calebe.MyGameObject;
+import com.calebe.MyOtherGameObject;
 import com.calebe.engine.game.GameObject;
 import dagger.Binds;
 import dagger.Module;
@@ -11,4 +12,8 @@ public abstract class GameObjectModule {
     @Binds
     @IntoSet
     abstract GameObject myGameObject(MyGameObject myGameObject);
+
+    @Binds
+    @IntoSet
+    abstract GameObject otherGameObject(MyOtherGameObject otherGameObject);
 }